Code cleanup
This is mostly dozen of small changes to eliminate MatLab warnings.  In most cases this means conforming better to standard MatLab conventions.  In others, unsed variables could be eliminated.  Some warnings were not applicable so I disabled them on the affected lines.
There is also some restructuring in A_Coral_Model*, also with the goal of cleaner more readable code.