Don't start workers for one thread
Three in one.
A bug fix: for small numbers of reefs (large everyx), too many workers could be specified to the parfor.
A development feature: by setting threads to zero debugging can be used without manually editing out the parfor loop call.  Unfortunately this works in called functions but not in the main program.
An optimization: By turning off pool creating and the overhead of moving code and data to a worker thread, single-core operation is up to twice as fast on the first run and 10% or so thereafter.