Fix supersymbiont seeding
Supersymbionts were getting seeded after the start year even if the advantage was set to zero.  The actual effect is minimal (not a single digit change in the end-of-run full reef table rows or 1985-2010 bleaching), but it's wrong.  Now to start year is set to the end of the run unless there's a actual use of super symbionts.