Clean up turning the various plots and maps on and off
The accumulated figure types often required commenting blocks of code in or out between runs.  Now all of the major figures can be controlled with flags (names starting with "do') at the top of the file.  This should be less time consuming and error prone.
The main file is almost 80 lines shorter now thanks to moving code out to functions.  It's still too long!