Removed wasted code
In initialization we were building from scratch an array of time values called "tim".  It was just a duplicate of TIME, so now tim is gone.  Also, there was some manipulation of SST values just to get the number of months in the run.   Now just use the length of TIME.
Also removed obsolete comments.