Move independent input calculations out of ODE function
Removed ri, gi and vgi computations from the ode function since they are independent of the solution.  Now the solver has fewer variables to deal with.  Also replaced "interp1" which turned out to be the biggest time-sink with "interp1q".  Results are the same for linear 1D interpolation.

This is still not iterating over multiple reefs or including OA and super symbionts.  It's all possible, but I want to check accuracy on one reef first.

Now only 2.2 times slower - the ratio may change after compilation.