Optimization and psw2 updates, and Dormand-Prince minutia
Main program: A few changes to more easily turn stats for the stress plots on and off.  Put Dormand-Prince on hold fro now.  Add title option to 4-panel stress plot.  Change getPropTest to use switch/case instead of if/elseif and update with new options for 3% and 10% bleaching targets.   Optimize_psw2.mat now has the new values.  The ones we were using recently (20-27) are still in place with the same indexes, but the new values are now in slots 2-17.  OptimizePropCalc now verifies that the RCP specified matches the model runs and the SST loading.  Stats_Tables no longer needs to be edited when changing the table columns for plotting or optimization.  It's all controlled by arguments detailStats and optimizerMode.   tryDormandPrince now has the simple code to allow setting the accuracy level.