Changes to model constants!
Coral_Sym_constants_3 replaces _2.  The massive coral seed is now 7.4125E7 which is 6.25E6 * 11.86 as described in Baskett 2009.  The paper seems to have a typo, showing 7.7412E7.

Massive and branching coral seeds are now calculated based on 1% and 0.1% of their respective K values.  Symbiont seeds are now 10E-5% of their carrying capacity when coral is down to seed values.  In other words, 10E-7 * KS * C_seed.

Other changes in the main program and other *.m files upated here are just the values which happened to be used in recent runs, but I'm incuding them to be sure it's all up to date.